+++ jaguar_jwt/CHANGELOG.md
@@ -5,6 +5,7 @@
 + Added support for optional Not Before (`nbf`) time claims.
 + Fixed validation to reject token when current time equals the Expiry time.
 + Added support for custom payload name.
++ Added more validation unit tests.
 
 ## 2.1.2